requirements.txt
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
decorator==4.4.2
defusedxml==0.7.1
distro==1.9.0
google-auth==2.46.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==1.0.3
numpy==2.4.0
pillow==11.3.0
proglog==0.1.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==15.0.1
youtube-transcript-api==1.2.3
yt-dlp==2025.12.8
cloudinary==1.44.1